requirements.txt
asgiref==3.2.7
astroid==2.4.1
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.6
django-braces==1.14.0
django-embed-video==1.3.2
django-memcache-status==2.2
djangorestframework==3.12.1
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.2
python-memcached==1.59
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
